+++ addons/mail/static/tests/document_thread_window_tests.js
@@ -1,4 +1,4 @@
-odoo.define('mail.document_thread_window_tests', function (require) {
+odoo.define('mail.documentThreadWindowTests', function (require) {
 "use strict";
 
 var mailTestUtils = require('mail.testUtils');
@@ -7,8 +7,8 @@ var MessagingMenu = require('mail.systray.MessagingMenu');
 var testUtils = require('web.test_utils');
 
 QUnit.module('mail', {}, function () {
-
-QUnit.module('ThreadWindow (Document Thread)', {
+QUnit.module('Thread Window', {}, function () {
+QUnit.module('Document Thread', {
     beforeEach: function () {
         var partnerID = 44;
         this.data = {
@@ -277,5 +277,5 @@ QUnit.test('open, fold, unfold and close a document thread window', function (as
 });
 
 });
-
+});
 });
